Fix type for resolution fields.
git-svn-id: http://www.telegraphics.com.au/svn/psdparse/trunk@411 d127708d-35f1-0310-882e-b6c776866be2

+++ main.c
@@ -29,7 +29,7 @@ int verbose = DEFAULT_VERBOSE, quiet = 0, rsrc = 0, print_rsrc = 0, resdump = 0,
 	scavenge = 0, scavenge_psb = 0, scavenge_depth = 8, scavenge_mode = -1,
 	scavenge_rows = 0, scavenge_cols = 0, scavenge_chan = 3, scavenge_rle = 0,
 	makedirs = 0, numbered = 0, help = 0, split = 0, xmlout = 0, unicode_filenames = 0;
-long hres, vres; // we don't use these, but they're set within doresources()
+uint32_t hres, vres; // we don't use these, but they're set within doresources()
 
 #ifdef ALWAYS_WRITE_PNG
 	// for the Windows console app, we want to be able to drag and drop a PSD
